Modified when the slide in animation occurs
To prevent the fade-slide-in animation from happening on page load, it
will now enable itself on the first page change. This is to prevent the
user from seeing the fully pre-rendered page initially, but then
immediately re-render with the animation. It's a bit of a hacky
solution, but it's been working on great in my testing.