Added data persistence for fetched data
For the pages that fetch data from the API (BlogListPage, BlogEntryPage,
and FavoritesOf), the data fetched during the prerender phase will now
be perisisted on the client side instead of being fetched again. This
should help out with both the loading screen flashing on these pages
(Even though it appeared to have rendered the data before) and with
unnecessary data fetching (Fetched on the server and client side).